Let the admin add house rules to the extraction prompt, and give Settings tabs
Two changes, in this order, because the second is what made the first urgent.

## Extra extraction rules

The extraction system prompt was a fixed string literal. The only part of it an
admin could vary was the result language; everything else -- how to pick a
title, which correspondent to prefer, what to tag -- was baked in, so an
operator who wanted "treat Rechnung as the document type Invoice" or "tag
insurance documents with the policy number" had to fork.

app_settings gains extraction_rules: up to 4000 characters, appended to the
built-in prompt. Empty, the default, leaves the prompt byte-for-byte what it
was.

It is an addition, not a replacement. The prompt's field list is coupled to
models.ExtractedMetadata, and the correspondent/document-type catalog blocks
are what make reuse and dedup work, so both stay. The rules land after those
catalog blocks and *before* the closing document_date / no-markdown block: the
format instructions a model follows best are the ones it read last, and a house
rule is not entitled to loosen the JSON contract. The block says as much --
"never add, rename or drop fields because of them" -- and, unlike the catalog,
is not labelled untrusted: only an admin can write it, through a route behind
bindAdmin.

The rules are tenant-owned, like processing_result_language and
always_require_review, so they stay out of touchesManaged() and a managed
instance keeps them. No environment variable: this is a Settings choice, not a
boot secret. The runtime already rebuilds the extractor from the app_settings
record hooks, so a save applies without a restart.

NewExtractor took the new argument; NewOpenAIClient did not. That constructor is
shared with chat, search, the splitter and LLM OCR -- five callers and some
twenty test call sites that already pass two empty strings for extraction-only
fields -- and NewExtractor has exactly one caller. It sets the field on the
client it builds instead.

ExtractMetadata logs the rules' length as rule_chars, so "why did this document
come out like that" has an answer in the log without putting the operator's own
text there.

## Settings tabs

frontend/src/routes/settings.tsx was 966 lines rendering six sections stacked on
one screen, all sharing one <form id="app-settings"> with five Save buttons
attached to it by id -- each of which saved every field on the page. One more
field made that worse, not better.

Settings is now a shell with one tab per section, on real paths, which is what
/upload and /import already do:

  /settings             Appearance
  /settings/ai          Providers + Models (one subject, hidden together on a
                        managed instance)
  /settings/processing   timeouts, languages, extraction rules, review policy
  /settings/worker       job timeout, retries
  /settings/duplicates   detection and threshold

Appearance sits on /settings itself, the way Files sits on /upload, so every
existing link to /settings still lands somewhere sensible.

A form can no longer span sibling routes -- and the AI tab has a provider form
of its own, which would have nested -- so each tab owns its form and PATCHes
only its own fields. Every field in settingsPatchRequest is already a pointer
where nil means "keep what is stored", so this needed no API change and deleted
two things: FORM_ID with its scattered form={FORM_ID} attributes, and the
...(aiEditable ? {...} : {}) spread that kept a managed instance from naming
fields it would be refused. The managed-only fields now live in a tab a managed
instance does not render.

One behaviour change worth stating: Save under Processing no longer writes your
edits on the Worker tab. That is the readable behaviour; the old Save quietly
wrote five sections at once.

The shared plumbing is one hook, useSettingsForm, built on the existing useAsync
-- no React context, of which this codebase has none. It holds the tab's slice
of the record, the save, and the feedback state; ResultDialog and
SaveSettingsButton moved into components/settings/ beside it. Each tab kept its
own validation along with its own fields.

The providers list and its add/edit drawer moved to
components/settings/ProvidersBlock.tsx so no file is over ~280 lines; the
biggest thing here was 966.

tabClassName was already copy-pasted in routes/upload.tsx and routes/import.tsx.
A third copy is where it belongs in components/ui.tsx, so all three now import
it.

## Always require review

Moved to the top of the Processing tab, in a bordered panel of its own with a
larger accented checkbox. It is the only setting on that tab that decides where
a finished document goes, and as small muted text at the foot of a two-column
grid it was the easiest one on the page to miss.

## Notes

Migration 1730000029 adds the field, guarded and re-runnable: a managed instance
re-runs every migration on every boot. The length cap is checked in
applySettingsPatch as well as by the field's Max, so the 400 carries a message
an admin can read.

e2e for both halves is on the matching lemmary-dev branch, pushed first.
