+++ frontend/src/router.test.ts
@@ -22,6 +22,10 @@ describe('route tree', () => {
       '/rag/research/$sessionId',
       '/ocr-test',
       '/settings',
+      '/settings/ai',
+      '/settings/processing',
+      '/settings/worker',
+      '/settings/duplicates',
       '/management',
       '/import',
       '/import/ngx',
