+++ frontend/src/routes/settings.tsx
@@ -1,429 +1,27 @@
-import { type SubmitEvent, useCallback, useEffect, useRef, useState } from 'react'
-import { Link } from '@tanstack/react-router'
-import {
-  createAIProvider,
-  deleteAIProvider,
-  listAIProviders,
-  requiresAPIKey,
-  requiresSignIn,
-  sdkAliasDefault,
-  sdkLabel,
-  updateAIProvider,
-  keylessProviderDocs,
-  keylessProviderHint,
-  SDK_DEFAULT_BASE,
-  SDK_OPTIONS,
-  type AIProvider,
-  type ProviderSDK,
-} from '../lib/api/providers'
-import { ChatGPTSignIn } from '../components/ChatGPTSignIn'
-import {
-  getAppSettings,
-  getEmbeddingStats,
-  updateAppSettings,
-  type AppSettings,
-  type EmbeddingStats,
-} from '../lib/api/settings'
-import { ProviderModelFields } from '../components/ProviderModelFields'
-import { DEFAULT_ACCENT } from '../lib/api/meta'
+import { Link, Outlet } from '@tanstack/react-router'
+
 import { useAppMeta } from '../hooks/useAppMeta'
-import {
-  Button,
-  DocsLink,
-  inputClassName,
-  labelClassName,
-  fieldHintClassName,
-  labelTextClassName,
-  sectionClassName,
-  sectionTitleClassName,
-} from '../components/ui'
+import { tabClassName } from '../components/ui'
 
 /**
- * Appearance stands above Providers, which owns a form of its own, so its
- * fields sit outside the settings form and reach it by id instead: forms
- * cannot nest, and `form=` is how HTML attaches an input to one elsewhere on
- * the page.
+ * The Settings shell: a header, the tabs, and whichever tab is on the URL.
+ *
+ * The sections used to be stacked on one page sharing one form, which had grown
+ * past what anyone could scan. Each is now its own route, and each saves only
+ * its own fields -- the settings PATCH treats every field as optional, so a tab
+ * naming its own leaves the rest of the record alone.
+ *
+ * Admin access is enforced by the route's beforeLoad guard, which the tab routes
+ * inherit.
  */
-const FORM_ID = 'app-settings'
-
-function SaveSettingsButton({ saving }: { saving: boolean }) {
-  return (
-    <div>
-      <Button type="submit" form={FORM_ID} disabled={saving}>
-        {saving ? 'Saving...' : 'Save settings'}
-      </Button>
-    </div>
-  )
-}
-
-/**
- * Save feedback as a modal: the inline lines at the foot of a page this long
- * were read by nobody. Success clears itself, an error waits to be dismissed.
- */
-function ResultDialog({
-  error,
-  success,
-  onClose,
-}: {
-  error: string
-  success: string
-  onClose: () => void
-}) {
-  const ref = useRef<HTMLDialogElement>(null)
-  const open = Boolean(error || success)
-
-  useEffect(() => {
-    const dialog = ref.current
-    if (!dialog) return
-    if (!open) {
-      dialog.close()
-      return
-    }
-    dialog.showModal()
-    if (error) return
-    const timer = setTimeout(onClose, 1500)
-    return () => clearTimeout(timer)
-  }, [open, error, onClose])
-
-  return (
-    <dialog
-      ref={ref}
-      onClose={onClose}
-      className="m-auto max-w-sm border border-line bg-surface p-5 text-ink backdrop:bg-ink/40"
-    >
-      <p className={`text-sm ${error ? 'text-madder' : 'text-forest'}`}>{error || success}</p>
-      {error && (
-        <div className="mt-4 flex justify-end">
-          <Button onClick={onClose}>Close</Button>
-        </div>
-      )}
-    </dialog>
-  )
-}
-
-type FormState = {
-  ocr_provider_id: string
-  ocr_model: string
-  extract_provider_id: string
-  extract_model: string
-  chat_provider_id: string
-  chat_model: string
-  search_provider_id: string
-  search_model: string
-  search_helper_provider_id: string
-  search_helper_model: string
-  embedding_provider_id: string
-  embedding_model: string
-  ocr_timeout_sec: string
-  processing_result_language: string
-  deep_search_languages: string
-  openai_timeout_sec: string
-  worker_timeout_sec: string
-  worker_max_retries: string
-  always_require_review: boolean
-  near_duplicate_detection_enabled: boolean
-  near_duplicate_threshold: string
-  app_name: string
-  accent: string
-}
-
-function formFromSettings(settings: AppSettings): FormState {
-  return {
-    ocr_provider_id: settings.ocr_provider_id,
-    ocr_model: settings.ocr_model,
-    extract_provider_id: settings.extract_provider_id,
-    extract_model: settings.extract_model,
-    chat_provider_id: settings.chat_provider_id,
-    chat_model: settings.chat_model,
-    search_provider_id: settings.search_provider_id,
-    search_model: settings.search_model,
-    search_helper_provider_id: settings.search_helper_provider_id,
-    search_helper_model: settings.search_helper_model,
-    embedding_provider_id: settings.embedding_provider_id,
-    embedding_model: settings.embedding_model,
-    ocr_timeout_sec: String(settings.ocr_timeout_sec),
-    processing_result_language: settings.processing_result_language,
-    deep_search_languages: settings.deep_search_languages,
-    openai_timeout_sec: String(settings.openai_timeout_sec),
-    worker_timeout_sec: String(settings.worker_timeout_sec),
-    worker_max_retries: String(settings.worker_max_retries),
-    always_require_review: settings.always_require_review,
-    near_duplicate_detection_enabled: settings.near_duplicate_detection_enabled,
-    near_duplicate_threshold: String(settings.near_duplicate_threshold ?? 0.92),
-    app_name: settings.app_name,
-    // The color input has no empty state, so it shows the accent actually in
-    // force. Saving an untouched form therefore writes the default down, which
-    // is what it was already resolving to.
-    accent: settings.accent || DEFAULT_ACCENT,
-  }
-}
-
-// The four LLM bindings the simple view collapses into one "general" model.
-// An empty Deep Search helper still fits: it means the search model does that
-// work itself, which is the same model either way.
-function fitsOneGeneralModel(form: FormState): boolean {
-  const general = `${form.extract_provider_id}|${form.extract_model}`
-  const helper = `${form.search_helper_provider_id}|${form.search_helper_model}`
-  return (
-    `${form.chat_provider_id}|${form.chat_model}` === general &&
-    `${form.search_provider_id}|${form.search_model}` === general &&
-    (helper === general || helper === '|')
-  )
-}
-
-type ProviderDraft = {
-  sdk: ProviderSDK
-  alias: string
-  base_url: string
-  api_key: string
-}
-
-function emptyDraft(sdk: ProviderSDK = 'openai'): ProviderDraft {
-  return { sdk, alias: '', base_url: SDK_DEFAULT_BASE[sdk], api_key: '' }
-}
-
-// EmbeddingStatsLine says how much of the archive the chosen model has actually
-// covered. Without it, switching the model looks instantaneous while the
-// backfill is in fact working through the archive a batch a minute, and nothing
-// on the page would say so.
-function EmbeddingStatsLine({ stats }: { stats: EmbeddingStats | null }) {
-  if (!stats || !stats.enabled) return null
-
-  const parts = [`${stats.embedded} of ${stats.total} documents embedded`]
-  if (stats.dims > 0) parts.push(`${stats.dims} dimensions`)
-  if (stats.chunks > 0) parts.push(`${stats.chunks} passages`)
-  if (stats.pending > 0) parts.push(`${stats.pending} queued`)
-  if (stats.failed > 0) parts.push(`${stats.failed} failed`)
-
-  return <p className={fieldHintClassName}>{parts.join(' · ')}.</p>
-}
-
-// Admin access is enforced by the route's beforeLoad guard.
 export function SettingsPage() {
   // unknown/failed meta counts as managed; see AppMeta.aiManaged
-  const { aiManaged, chatgptLogin } = useAppMeta()
+  const { aiManaged } = useAppMeta()
   const aiEditable = aiManaged === false
-  const [form, setForm] = useState<FormState | null>(null)
-  const [embeddingStats, setEmbeddingStats] = useState<EmbeddingStats | null>(null)
-  const [providers, setProviders] = useState<AIProvider[]>([])
-  const [draft, setDraft] = useState<ProviderDraft>(emptyDraft())
-  const [editingId, setEditingId] = useState<string | null>(null)
-  const [showAdd, setShowAdd] = useState(false)
-  const [loading, setLoading] = useState(true)
-  const [advancedModels, setAdvancedModels] = useState(false)
-  const [saving, setSaving] = useState(false)
-  const [error, setError] = useState('')
-  const [success, setSuccess] = useState('')
-
-  const closeResult = useCallback(() => {
-    setError('')
-    setSuccess('')
-  }, [])
-
-  async function reloadProviders() {
-    const next = await listAIProviders()
-    setProviders(next)
-    return next
-  }
-
-  useEffect(() => {
-    let active = true
-
-    async function load() {
-      try {
-        const [settings, nextProviders] = await Promise.all([getAppSettings(), listAIProviders()])
-        if (!active) return
-        setForm(formFromSettings(settings))
-        setProviders(nextProviders)
-        // Separately, and allowed to fail: it scans two tables, and a slow or
-        // broken count must not keep the Settings form off the screen.
-        getEmbeddingStats()
-          .then((stats) => {
-            if (active) setEmbeddingStats(stats)
-          })
-          .catch(() => {})
-        setError('')
-      } catch (err) {
-        if (active) {
-          setError(err instanceof Error ? err.message : 'Failed to load settings')
-        }
-      } finally {
-        if (active) setLoading(false)
-      }
-    }
-
-    void load()
-    return () => {
-      active = false
-    }
-  }, [])
-
-  // Takes a function as well as an object because the provider picker fires
-  // onProviderChange and onModelChange back to back: a patch built from the
-  // render's `form` would undo the first of the two.
-  function updateFields(patch: Partial<FormState> | ((current: FormState) => Partial<FormState>)) {
-    setForm((current) => {
-      if (!current) return current
-      return { ...current, ...(typeof patch === 'function' ? patch(current) : patch) }
-    })
-    setSuccess('')
-  }
-
-  function updateField<K extends keyof FormState>(key: K, value: FormState[K]) {
-    updateFields({ [key]: value } as Partial<FormState>)
-  }
-
-  /** The simple view has one LLM binding; the other three follow it. */
-  function setGeneralModel(providerId: string, model: string): Partial<FormState> {
-    return {
-      extract_provider_id: providerId,
-      extract_model: model,
-      chat_provider_id: providerId,
-      chat_model: model,
-      search_provider_id: providerId,
-      search_model: model,
-      // Named rather than left empty: empty means the same thing here, but a
-      // model spelled out is what Advanced setup then shows.
-      search_helper_provider_id: providerId,
-      search_helper_model: model,
-    }
-  }
-
-  async function onSaveProvider(event: SubmitEvent<HTMLFormElement>) {
-    event.preventDefault()
-    try {
-      setError('')
-      setSuccess('')
-      if (editingId) {
-        await updateAIProvider(editingId, {
-          sdk: draft.sdk,
-          alias: draft.alias.trim(),
-          base_url: draft.base_url.trim(),
-          ...(draft.api_key.trim() ? { api_key: draft.api_key.trim() } : {}),
-        })
-      } else {
-        await createAIProvider({
-          sdk: draft.sdk,
-          alias: draft.alias.trim() || sdkAliasDefault(draft.sdk),
-          base_url: draft.base_url.trim(),
-          api_key: draft.api_key.trim(),
-        })
-      }
-      await reloadProviders()
-      setDraft(emptyDraft())
-      setEditingId(null)
-      setShowAdd(false)
-      setSuccess('Provider saved.')
-    } catch (err) {
-      setError(err instanceof Error ? err.message : 'Failed to save provider')
-    }
-  }
-
-  async function onDeleteProvider(id: string) {
-    try {
-      setError('')
-      await deleteAIProvider(id)
-      await reloadProviders()
-      setSuccess('Provider deleted.')
-    } catch (err) {
-      setError(err instanceof Error ? err.message : 'Failed to delete provider')
-    }
-  }
-
-  async function onSubmit(event: SubmitEvent<HTMLFormElement>) {
-    event.preventDefault()
-    if (!form) return
-
-    const ocrTimeout = Number(form.ocr_timeout_sec)
-    const openAITimeout = Number(form.openai_timeout_sec)
-    const workerTimeout = Number(form.worker_timeout_sec)
-    const maxRetries = Number(form.worker_max_retries)
-    const nearThreshold = Number(form.near_duplicate_threshold)
-
-    if (!Number.isFinite(ocrTimeout) || ocrTimeout <= 0) {
-      setError('OCR timeout must be a positive number')
-      return
-    }
-    if (!Number.isFinite(openAITimeout) || openAITimeout <= 0) {
-      setError('AI timeout must be a positive number')
-      return
-    }
-    if (!Number.isFinite(workerTimeout) || workerTimeout <= 0) {
-      setError('Worker timeout must be a positive number')
-      return
-    }
-    if (!Number.isFinite(maxRetries) || maxRetries < 0) {
-      setError('Worker max retries must be >= 0')
-      return
-    }
-    if (!Number.isFinite(nearThreshold) || nearThreshold <= 0 || nearThreshold > 1) {
-      setError('Near-duplicate threshold must be between 0 and 1')
-      return
-    }
-    if (form.app_name.trim() === '') {
-      setError('Application name is required')
-      return
-    }
-    if (!/^#[0-9a-fA-F]{6}$/.test(form.accent)) {
-      setError('Accent color must be a hex value like #6e2620')
-      return
-    }
-
-    try {
-      setSaving(true)
-      setError('')
-      setSuccess('')
-
-      const settings = await updateAppSettings({
-        app_name: form.app_name.trim(),
-        accent: form.accent,
-        ocr_timeout_sec: ocrTimeout,
-        processing_result_language: form.processing_result_language,
-        deep_search_languages: form.deep_search_languages,
-        openai_timeout_sec: openAITimeout,
-        worker_timeout_sec: workerTimeout,
-        worker_max_retries: maxRetries,
-        always_require_review: form.always_require_review,
-        // Omit on a managed instance: naming these fails the whole patch.
-        ...(aiEditable
-          ? {
-              ocr_provider_id: form.ocr_provider_id,
-              ocr_model: form.ocr_model,
-              extract_provider_id: form.extract_provider_id,
-              extract_model: form.extract_model,
-              chat_provider_id: form.chat_provider_id,
-              chat_model: form.chat_model,
-              search_provider_id: form.search_provider_id,
-              search_model: form.search_model,
-              search_helper_provider_id: form.search_helper_provider_id,
-              search_helper_model: form.search_helper_model,
-              embedding_provider_id: form.embedding_provider_id,
-              embedding_model: form.embedding_model,
-              near_duplicate_detection_enabled: form.near_duplicate_detection_enabled,
-              near_duplicate_threshold: nearThreshold,
-            }
-          : {}),
-      })
-
-      setForm(formFromSettings(settings))
-      setSuccess('Settings saved. Runtime reloaded.')
-    } catch (err) {
-      setError(err instanceof Error ? err.message : 'Failed to save settings')
-    } finally {
-      setSaving(false)
-    }
-  }
-
-  if (loading || !form) {
-    return <p className="text-sm text-ink-soft">{error || 'Loading settings...'}</p>
-  }
-
-  const keylessDocs = keylessProviderDocs(draft.sdk)
-
 
   return (
     <div className="mx-auto max-w-3xl">
-      <div className="mb-6">
+      <div className="mb-5">
         <h1 className="font-display text-3xl font-semibold tracking-tight text-ink">Settings</h1>
         <p className="mt-1 text-sm text-ink-soft">
           Runtime configuration for OCR, AI, and the worker. Changes apply immediately.
@@ -435,532 +33,34 @@ export function SettingsPage() {
         )}
       </div>
 
-      <section className={`${sectionClassName} mb-5`}>
-        <h2 className={sectionTitleClassName}>Appearance</h2>
-        <div className="grid gap-4 sm:grid-cols-2">
-          <div className={labelClassName}>
-            <label className={labelClassName}>
-              <span className={labelTextClassName}>Application name</span>
-              <input
-                form={FORM_ID}
-                className={inputClassName}
-                value={form.app_name}
-                onChange={(e) => updateField('app_name', e.target.value)}
-              />
-            </label>
-            <p className={fieldHintClassName}>
-              Shown in the header and on the sign-in page, and used as the sender name on mail
-              and in passkey prompts.
-            </p>
-          </div>
-          <div className={labelClassName}>
-            <span className={labelTextClassName}>Accent color</span>
-            <div className="flex items-center gap-2">
-              <input
-                form={FORM_ID}
-                type="color"
-                aria-label="Accent color"
-                className="h-9 w-12 cursor-pointer border border-line bg-surface p-1"
-                value={form.accent}
-                onChange={(e) => updateField('accent', e.target.value)}
-              />
-              <input
-                form={FORM_ID}
-                className={`${inputClassName} font-mono`}
-                aria-label="Accent color hex"
-                spellCheck={false}
-                value={form.accent}
-                onChange={(e) => updateField('accent', e.target.value.trim())}
-              />
-              <Button
-                variant="secondary"
-                size="sm"
-                onClick={() => updateField('accent', DEFAULT_ACCENT)}
-              >
-                Reset
-              </Button>
-            </div>
-            <p className={fieldHintClassName}>
-              Colors the logo mark and the accents around it. Pick a swatch or paste a hex value
-              such as {DEFAULT_ACCENT}.
-            </p>
-          </div>
-        </div>
-        <div className="mt-4">
-          <SaveSettingsButton saving={saving} />
-        </div>
-      </section>
-
-      {aiEditable && (
-        <section className={`${sectionClassName} mb-5`}>
-          <h2 className={sectionTitleClassName}>Providers</h2>
-          <ul className="mb-4 flex flex-col gap-2">
-            {providers.length === 0 && <li className="text-sm text-ink-soft">No providers yet.</li>}
-            {providers.map((item) => (
-              <li
-                key={item.id}
-                className="flex flex-wrap items-center justify-between gap-2 rounded-xs border border-line bg-bright px-3 py-2"
-              >
-                <div>
-                  <p className="text-sm font-medium text-ink">{item.alias}</p>
-                  <p className="text-xs text-ink-soft">
-                    {sdkLabel(item.sdk)}
-                    {item.base_url ? ` · ${item.base_url}` : ''}
-                    {requiresSignIn(item.sdk)
-                      ? item.signed_in
-                        ? ' · signed in'
-                        : ' · not signed in'
-                      : requiresAPIKey(item.sdk)
-                        ? item.api_key_set
-                          ? ' · key set'
-                          : ' · missing key'
-                        : ' · no key needed'}
-                  </p>
-                  {/* The sign-in panel stands where the key field would, and on
-                      the saved row rather than in the add form: the flow needs
-                      a provider id to store the token against. */}
-                  {requiresSignIn(item.sdk) && (
-                    <ChatGPTSignIn provider={item} onChange={reloadProviders} />
-                  )}
-                </div>
-                <div className="flex gap-2">
-                  <Button
-                    variant="secondary"
-                    size="xs"
-                    onClick={() => {
-                      setEditingId(item.id)
-                      setShowAdd(true)
-                      setDraft({
-                        sdk: item.sdk,
-                        alias: item.alias,
-                        base_url: item.base_url,
-                        api_key: '',
-                      })
-                    }}
-                  >
-                    Edit
-                  </Button>
-                  <Button
-                    variant="secondary"
-                    size="xs"
-                    className="text-madder hover:bg-madder/10"
-                    onClick={() => void onDeleteProvider(item.id)}
-                  >
-                    Delete
-                  </Button>
-                </div>
-              </li>
-            ))}
-          </ul>
-          {!showAdd ? (
-            <Button
-              variant="secondary"
-              size="sm"
-              onClick={() => {
-                setEditingId(null)
-                setDraft(emptyDraft())
-                setShowAdd(true)
-              }}
-            >
-              Add provider
-            </Button>
-          ) : (
-            <form className="grid gap-3 sm:grid-cols-2" onSubmit={onSaveProvider}>
-              <label className={labelClassName}>
-                <span className={labelTextClassName}>SDK</span>
-                <select
-                  className={inputClassName}
-                  value={draft.sdk}
-                  onChange={(event) => {
-                    const sdk = event.target.value as ProviderSDK
-                    setDraft((current) => ({
-                      ...current,
-                      sdk,
-                      base_url:
-                        current.base_url === SDK_DEFAULT_BASE[current.sdk]
-                          ? SDK_DEFAULT_BASE[sdk]
-                          : current.base_url,
-                      // The key field is about to disappear; a value typed
-                      // before the switch would otherwise be posted invisibly.
-                      api_key: requiresAPIKey(sdk) ? current.api_key : '',
-                    }))
-                  }}
-                >
-                  {/* Offering an SDK the server will refuse is a dead end an
-                      admin cannot diagnose, so chatgpt appears only where
-                      AI_CHATGPT_LOGIN turned it on. */}
-                  {SDK_OPTIONS.filter(
-                    (option) => option.value !== 'chatgpt' || chatgptLogin === true,
-                  ).map((option) => (
-                    <option key={option.value} value={option.value}>
-                      {option.label}
-                    </option>
-                  ))}
-                </select>
-              </label>
-              <label className={labelClassName}>
-                <span className={labelTextClassName}>Alias</span>
-                <input
-                  className={inputClassName}
-                  value={draft.alias}
-                  placeholder={sdkAliasDefault(draft.sdk)}
-                  onChange={(event) => setDraft((current) => ({ ...current, alias: event.target.value }))}
-                />
-              </label>
-              {draft.sdk !== 'google_vision' && (
-                <label className={`${labelClassName} sm:col-span-2`}>
-                  <span className={labelTextClassName}>Base URL</span>
-                  <input
-                    className={inputClassName}
-                    value={draft.base_url}
-                    onChange={(event) =>
-                      setDraft((current) => ({ ...current, base_url: event.target.value }))
-                    }
-                  />
-                </label>
-              )}
-              {requiresAPIKey(draft.sdk) ? (
-                <label className={`${labelClassName} sm:col-span-2`}>
-                  <span className={labelTextClassName}>
-                    API key{editingId ? ' (leave blank to keep)' : ''}
-                  </span>
-                  <input
-                    type="password"
-                    autoComplete="off"
-                    className={inputClassName}
-                    value={draft.api_key}
-                    required={!editingId}
-                    onChange={(event) =>
-                      setDraft((current) => ({ ...current, api_key: event.target.value }))
-                    }
-                  />
-                </label>
-              ) : requiresSignIn(draft.sdk) ? (
-                <p className={`${fieldHintClassName} sm:col-span-2`}>
-                  Save the provider first, then sign in to it from the list above. Chat,
-                  extraction and Deep Search can run on the subscription; embeddings and OCR
-                  cannot, and keep whichever provider they have.
-                </p>
-              ) : (
-                <p className={`${fieldHintClassName} sm:col-span-2`}>
-                  {keylessProviderHint(draft.sdk)}{' '}
-                  {keylessDocs && (
-                    <DocsLink href={keylessDocs.href}>Read the {keylessDocs.label} guide.</DocsLink>
-                  )}
-                </p>
-              )}
-              <div className="flex gap-2 sm:col-span-2">
-                <Button type="submit" size="sm">
-                  {editingId ? 'Update provider' : 'Save provider'}
-                </Button>
-                <Button
-                  variant="secondary"
-                  size="sm"
-                  onClick={() => {
-                    setShowAdd(false)
-                    setEditingId(null)
-                  }}
-                >
-                  Cancel
-                </Button>
-              </div>
-            </form>
-          )}
-        </section>
-      )}
-
-      <form id={FORM_ID} className="flex flex-col gap-5" onSubmit={onSubmit}>
+      <nav
+        aria-label="Settings sections"
+        className="mb-5 flex flex-wrap items-center gap-5 border-b border-line"
+      >
+        <Link to="/settings" activeOptions={{ exact: true }} className={tabClassName}>
+          Appearance
+        </Link>
+        {/* Hidden rather than disabled on a managed instance: there is nothing
+            editable behind them. Typed in by hand they explain themselves. */}
         {aiEditable && (
-          <section className={sectionClassName}>
-            <h2 className={sectionTitleClassName}>Models</h2>
-            <div className="mb-4 flex items-center justify-between gap-3">
-              <p className={fieldHintClassName}>
-                {advancedModels
-                  ? 'Every job has its own model.'
-                  : fitsOneGeneralModel(form)
-                    ? 'One model does extraction, chat and search.'
-                    : 'Extraction, chat and search are on different models: open Advanced setup to see them. Picking a model here puts all three on it.'}
-              </p>
-              <Button
-                variant="secondary"
-                size="xs"
-                onClick={() => {
-                  // Collapsing on the way in, not on save: the simple view
-                  // shows one model, so the three it hides must already agree.
-                  if (advancedModels) {
-                    updateFields((current) =>
-                      setGeneralModel(current.extract_provider_id, current.extract_model),
-                    )
-                  }
-                  setAdvancedModels(!advancedModels)
-                }}
-              >
-                {advancedModels ? 'Simple setup' : 'Advanced setup'}
-              </Button>
-            </div>
-            <div className="grid gap-4 sm:grid-cols-2">
-              <ProviderModelFields
-                label="OCR"
-                help="Reads the text out of uploaded PDFs, images and scans. Plain text, CSV, Word and Excel files are read locally and skip this step."
-                providers={providers}
-                providerId={form.ocr_provider_id}
-                model={form.ocr_model}
-                purpose="ocr"
-                onProviderChange={(id) => updateField('ocr_provider_id', id)}
-                onModelChange={(value) => updateField('ocr_model', value)}
-              />
-              {!advancedModels && (
-                <ProviderModelFields
-                  label="General AI"
-                  help="Reads documents into metadata, answers questions on Ask AI, and runs Deep Search. Advanced setup splits this into one model per job."
-                  providers={providers}
-                  providerId={form.extract_provider_id}
-                  model={form.extract_model}
-                  purpose="llm"
-                  onProviderChange={(id) =>
-                    updateFields((current) => setGeneralModel(id, current.extract_model))
-                  }
-                  onModelChange={(value) =>
-                    updateFields((current) => setGeneralModel(current.extract_provider_id, value))
-                  }
-                />
-              )}
-
-              {advancedModels && (
-                <>
-                  <ProviderModelFields
-                    label="Extraction"
-                    help="Turns a document's text into its title, date, type, correspondent, tags and summary. Also proposes the cuts for Detect automatically when splitting a PDF."
-                    providers={providers}
-                    providerId={form.extract_provider_id}
-                    model={form.extract_model}
-                    purpose="llm"
-                    onProviderChange={(id) => updateField('extract_provider_id', id)}
-                    onModelChange={(value) => updateField('extract_model', value)}
-                  />
-                  <ProviderModelFields
-                    label="Chat"
-                    help="Answers questions about a single document on its Ask AI page. Leave the provider empty to turn the feature off."
-                    providers={providers}
-                    providerId={form.chat_provider_id}
-                    model={form.chat_model}
-                    purpose="llm"
-                    allowEmpty
-                    onProviderChange={(id) => updateField('chat_provider_id', id)}
-                    onModelChange={(value) => updateField('chat_model', value)}
-                  />
-                  <ProviderModelFields
-                    label="Search"
-                    help="Answers natural-language queries on the Deep Search page, in both Search and Research mode. Leave the provider empty to turn the feature off."
-                    providers={providers}
-                    providerId={form.search_provider_id}
-                    model={form.search_model}
-                    purpose="llm"
-                    allowEmpty
-                    onProviderChange={(id) => updateField('search_provider_id', id)}
-                    onModelChange={(value) => updateField('search_model', value)}
-                  />
-                  <ProviderModelFields
-                    label="Deep Search helper"
-                    help="Cheaper model Deep Search uses to read and extract from many documents at once: it turns long reads into notes and surveys whole topics one document at a time. Leave empty to have the Search model do this work itself."
-                    providers={providers}
-                    providerId={form.search_helper_provider_id}
-                    model={form.search_helper_model}
-                    purpose="llm"
-                    allowEmpty
-                    onProviderChange={(id) => updateField('search_helper_provider_id', id)}
-                    onModelChange={(value) => updateField('search_helper_model', value)}
-                  />
-                </>
-              )}
-
-              <ProviderModelFields
-                label="Embeddings"
-                help="Lets Deep Search find documents by meaning as well as by keyword, which is what makes a question phrased in one language reach a document written in another. Leave the provider empty to search by keyword only."
-                providers={providers}
-                providerId={form.embedding_provider_id}
-                model={form.embedding_model}
-                purpose="embedding"
-                allowEmpty
-                onProviderChange={(id) => updateField('embedding_provider_id', id)}
-                onModelChange={(value) => updateField('embedding_model', value)}
-              />
-              <div className="sm:col-span-2">
-                <EmbeddingStatsLine stats={embeddingStats} />
-              </div>
-            </div>
-          <div className="mt-4">
-            <SaveSettingsButton saving={saving} />
-          </div>
-          </section>
+          <Link to="/settings/ai" className={tabClassName}>
+            AI
+          </Link>
         )}
-
-        {/*
-          Kept out of Models on purpose: none of these is a provider or a model,
-          so a managed tenant keeps them. The environment seeds them on the first
-          boot and never touches them again, in either mode.
-        */}
-        <section className={sectionClassName}>
-          <h2 className={sectionTitleClassName}>Processing</h2>
-          <div className="grid gap-4 sm:grid-cols-2">
-            <div className={labelClassName}>
-              <label className={labelClassName}>
-                <span className={labelTextClassName}>OCR timeout (seconds)</span>
-                <input
-                  type="number"
-                  min={1}
-                  className={inputClassName}
-                  value={form.ocr_timeout_sec}
-                  onChange={(e) => updateField('ocr_timeout_sec', e.target.value)}
-                />
-              </label>
-              <p className={fieldHintClassName}>
-                How long one OCR call may take before the step fails.
-              </p>
-            </div>
-            <div className={labelClassName}>
-              <label className={labelClassName}>
-                <span className={labelTextClassName}>AI timeout (seconds)</span>
-                <input
-                  type="number"
-                  min={1}
-                  className={inputClassName}
-                  value={form.openai_timeout_sec}
-                  onChange={(e) => updateField('openai_timeout_sec', e.target.value)}
-                />
-              </label>
-              <p className={fieldHintClassName}>
-                How long one extraction, chat, search or split-detection request may take.
-              </p>
-            </div>
-            <div className={labelClassName}>
-              <label className={labelClassName}>
-                <span className={labelTextClassName}>Result language (ISO 639-1)</span>
-                <input
-                  className={inputClassName}
-                  placeholder="e.g. en"
-                  value={form.processing_result_language}
-                  onChange={(e) => updateField('processing_result_language', e.target.value)}
-                />
-              </label>
-              <p className={fieldHintClassName}>
-                Also stores the title, purpose, summary, type, correspondent and tags
-                translated into this language. Leave empty to keep only the document&rsquo;s own
-                language.
-              </p>
-            </div>
-            <div className={labelClassName}>
-              <label className={labelClassName}>
-                <span className={labelTextClassName}>Deep search languages</span>
-                <input
-                  className={inputClassName}
-                  placeholder="e.g. de,en,uk"
-                  value={form.deep_search_languages}
-                  onChange={(e) => updateField('deep_search_languages', e.target.value)}
-                />
-              </label>
-              <p className={fieldHintClassName}>
-                {form.embedding_model.trim() !== ''
-                  ? 'With an embedding model configured, one search already reaches documents in every language; this list is only used when Deep Search falls back to keyword search.'
-                  : 'Languages deep search translates keywords into, so a German invoice is found by an English question. Leave empty to search only in the language of the question.'}
-              </p>
-            </div>
-            <div className={`${labelClassName} sm:col-span-2`}>
-              <label className="flex items-center gap-2 text-sm text-ink-muted">
-                <input
-                  type="checkbox"
-                  checked={form.always_require_review}
-                  onChange={(e) => updateField('always_require_review', e.target.checked)}
-                />
-                Always require review for new documents
-              </label>
-              <p className={fieldHintClassName}>
-                Every document the AI extracts metadata for waits in the Inbox, however
-                confident the extraction was &mdash; reprocessed documents included. Nothing
-                completes but by your saying so. Off, only low-confidence extractions and
-                possible duplicates land there.
-              </p>
-            </div>
-          </div>
-          <div className="mt-4">
-            <SaveSettingsButton saving={saving} />
-          </div>
-        </section>
-
-        <section className={sectionClassName}>
-          <h2 className={sectionTitleClassName}>Worker</h2>
-          <div className="grid gap-4 sm:grid-cols-2">
-            <label className={labelClassName}>
-              <span className={labelTextClassName}>Job timeout (seconds)</span>
-              <input
-                type="number"
-                min={1}
-                className={inputClassName}
-                value={form.worker_timeout_sec}
-                onChange={(e) => updateField('worker_timeout_sec', e.target.value)}
-              />
-            </label>
-            <label className={labelClassName}>
-              <span className={labelTextClassName}>Max retries</span>
-              <input
-                type="number"
-                min={0}
-                className={inputClassName}
-                value={form.worker_max_retries}
-                onChange={(e) => updateField('worker_max_retries', e.target.value)}
-              />
-            </label>
-          </div>
-          <p className="mt-3 text-xs text-ink-soft">
-            Worker cron schedule stays in <code className="font-mono">WORKER_CRON_EXPR</code> in{' '}
-            <code className="font-mono">.env</code>.
-          </p>
-          <div className="mt-4">
-            <SaveSettingsButton saving={saving} />
-          </div>
-        </section>
-
+        <Link to="/settings/processing" className={tabClassName}>
+          Processing
+        </Link>
+        <Link to="/settings/worker" className={tabClassName}>
+          Worker
+        </Link>
         {aiEditable && (
-          <section className={sectionClassName}>
-            <h2 className={sectionTitleClassName}>Duplicates</h2>
-            <div className="grid gap-4 sm:grid-cols-2">
-              <label className="flex items-center gap-2 text-sm text-ink-muted sm:col-span-2">
-                <input
-                  type="checkbox"
-                  checked={form.near_duplicate_detection_enabled}
-                  onChange={(e) => updateField('near_duplicate_detection_enabled', e.target.checked)}
-                />
-                Enable near-duplicate detection after OCR (re-scans)
-              </label>
-              <label className={labelClassName}>
-                <span className={labelTextClassName}>Near-duplicate threshold (0–1)</span>
-                <input
-                  type="number"
-                  min={0.01}
-                  max={1}
-                  step={0.01}
-                  className={inputClassName}
-                  value={form.near_duplicate_threshold}
-                  onChange={(e) => updateField('near_duplicate_threshold', e.target.value)}
-                />
-              </label>
-            </div>
-            <p className="mt-3 text-xs text-ink-soft">
-              Exact file duplicates (same checksum) are always rejected on upload. Near-duplicate
-              matching compares OCR text and is off by default. Scan existing documents from{' '}
-              <Link to="/management" className="underline hover:text-oxblood">
-                Management
-              </Link>
-              .
-            </p>
-            <div className="mt-4">
-              <SaveSettingsButton saving={saving} />
-            </div>
-          </section>
+          <Link to="/settings/duplicates" className={tabClassName}>
+            Duplicates
+          </Link>
         )}
-      </form>
+      </nav>
 
-      <ResultDialog error={error} success={success} onClose={closeResult} />
+      <Outlet />
     </div>
   )
 }
