+++ frontend/src/routes/upload.tsx
@@ -3,9 +3,7 @@ import { Link, Outlet } from '@tanstack/react-router'
 
 import { LimitsUsage } from '../components/LimitsUsage'
 import { getLimits, type InstanceLimits } from '../lib/api/limits'
-
-const tabClassName =
-  '-mb-px border-b-2 border-transparent px-1 pb-2 pt-1 text-xs font-semibold uppercase tracking-[0.14em] text-ink-soft transition-colors hover:text-oxblood focus-visible:outline-2 focus-visible:outline-offset-4 focus-visible:outline-oxblood data-[status=active]:border-oxblood data-[status=active]:text-oxblood'
+import { tabClassName } from '../components/ui'
 
 export function UploadPage() {
   // Loaded on the shell rather than per tab, so the same figures show above
